Fix exception visibility Fix typos and comments
