+++ extensions/MongoDbAtlas/MongoDbAtlas.FunctionalTests/Startup.cs
@@ -1,4 +1,6 @@
-﻿/* IMPORTANT: the Startup class must be at the root of the namespace and
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+/* IMPORTANT: the Startup class must be at the root of the namespace and
  * the namespace must match exactly (required by Xunit.DependencyInjection) */
 
 namespace MongoDbAtlas.FunctionalTests;
