+++ extensions/MongoDbAtlas/MongoDbAtlas.FunctionalTests/Usings.cs
@@ -1,2 +1,4 @@
-﻿global using Xunit;
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+global using Xunit;
 global using Xunit.Abstractions;
